Move BAGL transaction UX code to separate file.
