+++ src/main.c
@@ -53,14 +53,6 @@ unsigned char G_io_seproxyhal_spi_buffer[IO_SEPROXYHAL_BUFFER_SIZE_B];
 #define COLOR_LIGHT_GRAY 0xEEEEEE
 
 #ifndef TARGET_FATSTACKS
-unsigned int io_seproxyhal_touch_verify_cancel(const bagl_element_t *e);
-unsigned int io_seproxyhal_touch_verify_ok(const bagl_element_t *e);
-unsigned int
-io_seproxyhal_touch_message_signature_verify_cancel(const bagl_element_t *e);
-unsigned int
-io_seproxyhal_touch_message_signature_verify_ok(const bagl_element_t *e);
-unsigned int io_seproxyhal_touch_display_token_cancel(const bagl_element_t *e);
-unsigned int io_seproxyhal_touch_display_token_ok(const bagl_element_t *e);
 unsigned int io_seproxyhal_touch_settings(const bagl_element_t *e);
 unsigned int io_seproxyhal_touch_exit(const bagl_element_t *e);
 #endif
@@ -80,598 +72,7 @@ const bagl_element_t *ui_menu_item_out_over(const bagl_element_t *e) {
     e = (const bagl_element_t *)(((unsigned int)e) + sizeof(bagl_element_t));
     return e;
 }
-#endif
 
-#if defined(TARGET_NANOS)
-
-const bagl_element_t ui_request_change_path_approval_nanos[] = {
-    UI_NANOS_BACKGROUND(),
-
-    UI_NANOS_TEXT(1, 0, 22, 128, "Warning !", BAGL_FONT_OPEN_SANS_LIGHT_16px),
-
-    UI_NANOS_TEXT(2, 0, 12, 128, "The change path", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_TEXT(2, 0, 26, 128, "is unusual", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(3, 0, 12, 128, "Change path", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(0x83, 15, 26, 98, vars.tmp_warning.derivation_path, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_ICON_LEFT(4, BAGL_GLYPH_ICON_CROSS),
-    UI_NANOS_ICON_RIGHT(4, BAGL_GLYPH_ICON_CHECK),
-    UI_NANOS_TEXT(4, 0, 12, 128, "Reject if you're", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_TEXT(4, 0, 26, 128, "not sure", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px)
-};
-
-unsigned int ui_request_change_path_approval_nanos_prepro(const bagl_element_t *element) {
-    if (element->component.userid > 0) {
-        unsigned int display = (ux_step == (0x7F & element->component.userid) - 1);
-        if (display) {
-            if (element->component.userid & 0x80) {
-                ux_loop_over_curr_element = 1;
-                UX_CALLBACK_SET_INTERVAL(MAX(
-                    3000, 1000 + bagl_label_roundtrip_duration_ms(element, 7)));
-            }
-        }
-        return display;
-    }
-    return 1;
-}
-
-unsigned int ui_request_change_path_approval_nanos_button(unsigned int button_mask,
-                                             unsigned int button_mask_counter);
-
-const bagl_element_t ui_verify_nanos[] = {
-    UI_NANOS_BACKGROUND(),
-    UI_NANOS_ICON_LEFT(0, BAGL_GLYPH_ICON_CROSS),
-    UI_NANOS_ICON_RIGHT(0, BAGL_GLYPH_ICON_CHECK),
-    UI_NANOS_TEXT(1, 0, 12, 128, "Confirm", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_TEXT(1, 0, 26, 128, "transaction", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(2, 0, 12, 128, "Amount", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(2, 23, 26, 82, vars.tmp.fullAmount, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(3, 0, 12, 128, "Address", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(3, 23, 26, 82, vars.tmp.fullAddress, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(4, 0, 12, 128, "Fees", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(4, 23, 26, 82, vars.tmp.feesAmount, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px)
-
-};
-unsigned int ui_verify_nanos_button(unsigned int button_mask,
-                                    unsigned int button_mask_counter);
-
-const bagl_element_t ui_verify_output_nanos[] = {
-
-    UI_NANOS_BACKGROUND(),
-    UI_NANOS_ICON_LEFT(0, BAGL_GLYPH_ICON_CROSS),
-    UI_NANOS_ICON_RIGHT(0, BAGL_GLYPH_ICON_CHECK),
-    UI_NANOS_TEXT(1, 0, 12, 128, "Confirm", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_TEXT(1, 0, 26, 128, vars.tmp.feesAmount, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(2, 0, 12, 128, "Amount", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(2, 23, 26, 82, vars.tmp.fullAmount, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(3, 0, 12, 128, "Address", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(3, 23, 26, 82, vars.tmp.fullAddress, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px)
-};
-
-unsigned int ui_verify_output_nanos_button(unsigned int button_mask,
-                                           unsigned int button_mask_counter);
-
-const bagl_element_t ui_finalize_nanos[] = {
-    UI_NANOS_BACKGROUND(),
-    UI_NANOS_ICON_LEFT(0, BAGL_GLYPH_ICON_CROSS),
-    UI_NANOS_ICON_RIGHT(0, BAGL_GLYPH_ICON_CHECK),
-    UI_NANOS_TEXT(1, 0, 12, 128, "Confirm", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_TEXT(1, 0, 26, 128, "transaction", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(2, 0, 12, 128, "Fees", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_SCROLLING_TEXT(2, 23, 26, 82, vars.tmp.feesAmount, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px)
-
-
-    /* TODO
-    {{BAGL_LABELINE                       , 0x02,   0,  12, 128,  12, 0, 0, 0 ,
-    COLOR_WHITE, COLOR_BLACK,
-    BAGL_FONT_OPEN_SANS_REGULAR_11px|BAGL_FONT_ALIGNMENT_CENTER, 0  }, "Amount",
-    0, 0, 0, NULL, NULL, NULL },
-    {{BAGL_LABELINE                       , 0x02,  23,  26,  82,  12, 0x80|10,
-    0, 0        , COLOR_WHITE, COLOR_BLACK,
-    BAGL_FONT_OPEN_SANS_EXTRABOLD_11px|BAGL_FONT_ALIGNMENT_CENTER, 26  },
-    vars.tmp.fullAmount, 0, 0, 0, NULL, NULL, NULL },
-    */
-};
-unsigned int ui_finalize_nanos_button(unsigned int button_mask,
-                                      unsigned int button_mask_counter);
-
-// display or not according to step, and adjust delay
-unsigned int ui_verify_prepro(const bagl_element_t *element) {
-    if (element->component.userid > 0) {
-        unsigned int display = (ux_step == element->component.userid - 1);
-        if (display) {
-            switch (element->component.userid) {
-            case 1:
-                UX_CALLBACK_SET_INTERVAL(2000);
-                break;
-            case 2:
-            case 3:
-            case 4:
-                UX_CALLBACK_SET_INTERVAL(MAX(
-                    3000, 1000 + bagl_label_roundtrip_duration_ms(element, 7)));
-                break;
-            }
-        }
-        return display;
-    }
-    return 1;
-}
-
-unsigned int ui_verify_output_prepro(const bagl_element_t *element) {
-    if (element->component.userid > 0) {
-        unsigned int display = (ux_step == element->component.userid - 1);
-        if (display) {
-            switch (element->component.userid) {
-            case 1:
-                UX_CALLBACK_SET_INTERVAL(2000);
-                break;
-            case 2:
-            case 3:
-                UX_CALLBACK_SET_INTERVAL(MAX(
-                    3000, 1000 + bagl_label_roundtrip_duration_ms(element, 7)));
-                break;
-            }
-        }
-        return display;
-    }
-    return 1;
-}
-
-unsigned int ui_finalize_prepro(const bagl_element_t *element) {
-    if (element->component.userid > 0) {
-        unsigned int display = (ux_step == element->component.userid - 1);
-        if (display) {
-            switch (element->component.userid) {
-            case 1:
-                UX_CALLBACK_SET_INTERVAL(2000);
-                break;
-            case 2:
-                UX_CALLBACK_SET_INTERVAL(MAX(
-                    3000, 1000 + bagl_label_roundtrip_duration_ms(element, 7)));
-                break;
-            }
-        }
-        return display;
-    }
-    return 1;
-}
-
-const bagl_element_t ui_verify_message_signature_nanos[] = {
-    UI_NANOS_BACKGROUND(),
-    UI_NANOS_ICON_LEFT(0, BAGL_GLYPH_ICON_CROSS),
-    UI_NANOS_ICON_RIGHT(0, BAGL_GLYPH_ICON_CHECK),
-    UI_NANOS_TEXT(1, 0, 12, 128, "Sign the", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-    UI_NANOS_TEXT(1, 0, 26, 128, "message", BAGL_FONT_OPEN_SANS_EXTRABOLD_11px),
-
-    UI_NANOS_TEXT(2, 0, 12, 128, "Message hash", BAGL_FONT_OPEN_SANS_REGULAR_11px),
-    UI_NANOS_SCROLLING_TEXT(2, 23, 26, 82, vars.tmp.fullAddress, BAGL_FONT_OPEN_SANS_EXTRABOLD_11px)
-};
-unsigned int
-ui_verify_message_signature_nanos_button(unsigned int button_mask,
-                                         unsigned int button_mask_counter);
-
-unsigned int ui_verify_message_prepro(const bagl_element_t *element) {
-    if (element->component.userid > 0) {
-        unsigned int display = (ux_step == element->component.userid - 1);
-        if (display) {
-            switch (element->component.userid) {
-            case 1:
-                UX_CALLBACK_SET_INTERVAL(2000);
-                break;
-            case 2:
-                UX_CALLBACK_SET_INTERVAL(MAX(
-                    3000, 1000 + bagl_label_roundtrip_duration_ms(element, 7)));
-                break;
-            }
-        }
-        return display;
-    }
-    return 1;
-}
-
-#endif // #if defined(TARGET_NANOS)
-#ifndef TARGET_FATSTACKS
-unsigned int io_seproxyhal_touch_verify_cancel(const bagl_element_t *e) {
-    // user denied the transaction, tell the USB side
-    if (!btchip_bagl_user_action(0)) {
-        // redraw ui
-        ui_idle();
-    }
-    return 0; // DO NOT REDRAW THE BUTTON
-}
-
-unsigned int io_seproxyhal_touch_verify_ok(const bagl_element_t *e) {
-    // user accepted the transaction, tell the USB side
-    if (!btchip_bagl_user_action(1)) {
-        // redraw ui
-        ui_idle();
-    }
-    return 0; // DO NOT REDRAW THE BUTTON
-}
-
-unsigned int
-io_seproxyhal_touch_message_signature_verify_cancel(const bagl_element_t *e) {
-    // user denied the transaction, tell the USB side
-    btchip_bagl_user_action_message_signing(0);
-    // redraw ui
-    ui_idle();
-    return 0; // DO NOT REDRAW THE BUTTON
-}
-
-unsigned int
-io_seproxyhal_touch_message_signature_verify_ok(const bagl_element_t *e) {
-    // user accepted the transaction, tell the USB side
-    btchip_bagl_user_action_message_signing(1);
-    // redraw ui
-    ui_idle();
-    return 0; // DO NOT REDRAW THE BUTTON
-}
-
-unsigned int io_seproxyhal_touch_display_token_cancel(const bagl_element_t *e) {
-    // revoke previous valid token if there was one
-    btchip_context_D.has_valid_token = false;
-    // user denied the token, tell the USB side
-    btchip_bagl_user_action_display(0);
-    // redraw ui
-    ui_idle();
-    return 0; // DO NOT REDRAW THE BUTTON
-}
-
- unsigned int io_seproxyhal_touch_display_token_ok(const bagl_element_t *e) {
-    // Set the valid token flag
-    btchip_context_D.has_valid_token = true;
-    // user approved the token, tell the USB side
-    btchip_bagl_user_action_display(1);
-    // redraw ui
-    ui_idle();
-    return 0; // DO NOT REDRAW THE BUTTON
-}
-#endif
-
-#if defined(TARGET_NANOS)
-unsigned int ui_verify_nanos_button(unsigned int button_mask,
-                                    unsigned int button_mask_counter) {
-    switch (button_mask) {
-    case BUTTON_EVT_RELEASED | BUTTON_LEFT:
-        io_seproxyhal_touch_verify_cancel(NULL);
-        break;
-
-    case BUTTON_EVT_RELEASED | BUTTON_RIGHT:
-        io_seproxyhal_touch_verify_ok(NULL);
-        break;
-    }
-    return 0;
-}
-
-unsigned int ui_verify_output_nanos_button(unsigned int button_mask,
-                                           unsigned int button_mask_counter) {
-    switch (button_mask) {
-    case BUTTON_EVT_RELEASED | BUTTON_LEFT:
-        io_seproxyhal_touch_verify_cancel(NULL);
-        break;
-
-    case BUTTON_EVT_RELEASED | BUTTON_RIGHT:
-        io_seproxyhal_touch_verify_ok(NULL);
-        break;
-    }
-    return 0;
-}
-
-unsigned int ui_finalize_nanos_button(unsigned int button_mask,
-                                      unsigned int button_mask_counter) {
-    switch (button_mask) {
-    case BUTTON_EVT_RELEASED | BUTTON_LEFT:
-        io_seproxyhal_touch_verify_cancel(NULL);
-        break;
-
-    case BUTTON_EVT_RELEASED | BUTTON_RIGHT:
-        io_seproxyhal_touch_verify_ok(NULL);
-        break;
-    }
-    return 0;
-}
-
-unsigned int
-ui_verify_message_signature_nanos_button(unsigned int button_mask,
-                                         unsigned int button_mask_counter) {
-    switch (button_mask) {
-    case BUTTON_EVT_RELEASED | BUTTON_LEFT:
-        io_seproxyhal_touch_message_signature_verify_cancel(NULL);
-        break;
-
-    case BUTTON_EVT_RELEASED | BUTTON_RIGHT:
-        io_seproxyhal_touch_message_signature_verify_ok(NULL);
-        break;
-    }
-    return 0;
-}
-
- unsigned int ui_request_change_path_approval_nanos_button(unsigned int button_mask,
-                                             unsigned int button_mask_counter)
-{
-    if (ux_step == 3)
-    {
-        switch (button_mask)
-        {
-        case BUTTON_EVT_RELEASED | BUTTON_LEFT:
-            io_seproxyhal_touch_display_cancel(NULL);
-            break;
-    case BUTTON_EVT_RELEASED | BUTTON_RIGHT:
-        io_seproxyhal_touch_display_ok(NULL);
-        break;
-    }
-    }
-    else
-    {
-        if(button_mask == (BUTTON_EVT_RELEASED | BUTTON_LEFT | BUTTON_RIGHT))
-        {
-                // if we were looping over a single element, disable this loop and diffuse the redisplay timeout (used by scrolling text)
-                if(ux_loop_over_curr_element) {
-                    ux_loop_over_curr_element = 0;
-                    ux.callback_interval_ms = 0;
-                }
-                // prepare next screen
-                ux_step = (ux_step + 1) % ux_step_count;
-                // redisplay screen
-                UX_REDISPLAY();
-        }
-    }
-    return 0;
-}
-
-#endif // #if defined(TARGET_NANOS)
-
-#if defined(TARGET_NANOX) || defined(TARGET_NANOS2)
-
-//////////////////////////////////////////////////////////////////////
-UX_STEP_NOCB(
-    ux_sign_flow_1_step,
-    pnn, 
-    {
-      &C_icon_certificate,
-      "Sign",
-      "message",
-    });
-UX_STEP_NOCB(
-    ux_sign_flow_2_step,
-    bnnn_paging, 
-    {
-      .title = "Message hash",
-      .text = vars.tmp.fullAddress,
-    });
-UX_STEP_VALID(
-    ux_sign_flow_3_step,
-    pbb,
-    io_seproxyhal_touch_message_signature_verify_ok(NULL),
-    {
-      &C_icon_validate_14,
-      "Sign",
-      "message",
-    });
-UX_STEP_VALID(
-    ux_sign_flow_4_step,
-    pbb,
-    io_seproxyhal_touch_message_signature_verify_cancel(NULL),
-    {
-      &C_icon_crossmark,
-      "Cancel",
-      "signature",
-    });
-
-UX_FLOW(ux_sign_flow,
-  &ux_sign_flow_1_step,
-  &ux_sign_flow_2_step,
-  &ux_sign_flow_3_step,
-  &ux_sign_flow_4_step
-);
-
-//////////////////////////////////////////////////////////////////////
-
-UX_STEP_NOCB(ux_confirm_full_flow_1_step,
-    pnn, 
-    {
-      &C_icon_eye,
-      "Review",
-      "transaction",
-    });
-UX_STEP_NOCB(
-    ux_confirm_full_flow_2_step,
-    bnnn_paging, 
-    {
-      .title = "Amount",
-      .text = vars.tmp.fullAmount
-    });
-UX_STEP_NOCB(
-    ux_confirm_full_flow_3_step,
-    bnnn_paging, 
-    {
-      .title = "Address",
-      .text = vars.tmp.fullAddress,
-    });
-UX_STEP_NOCB(
-    ux_confirm_full_flow_4_step,
-    bnnn_paging, 
-    {
-      .title = "Fees",
-      .text = vars.tmp.feesAmount,
-    });
-UX_STEP_VALID(
-    ux_confirm_full_flow_5_step,
-    pbb,
-    io_seproxyhal_touch_verify_ok(NULL),
-    {
-      &C_icon_validate_14,
-      "Accept",
-      "and send",
-    });
-UX_STEP_VALID(
-    ux_confirm_full_flow_6_step,
-    pb,
-    io_seproxyhal_touch_verify_cancel(NULL),
-    {
-      &C_icon_crossmark,
-      "Reject",
-    });
-// confirm_full: confirm transaction / Amount: fullAmount / Address: fullAddress / Fees: feesAmount
-UX_FLOW(ux_confirm_full_flow,
-  &ux_confirm_full_flow_1_step,
-  &ux_confirm_full_flow_2_step,
-  &ux_confirm_full_flow_3_step,
-  &ux_confirm_full_flow_4_step,
-  &ux_confirm_full_flow_5_step,
-  &ux_confirm_full_flow_6_step
-);
-
-//////////////////////////////////////////////////////////////////////
-
-UX_STEP_NOCB(
-    ux_confirm_single_flow_1_step, 
-    pnn, 
-    {
-      &C_icon_eye,
-      "Review",
-      "transaction"
-    });
-UX_STEP_NOCB(
-    ux_confirm_single_flow_2_step, 
-    bnnn_paging, 
-    {
-      .title = "Amount",
-      .text = vars.tmp.fullAmount,
-    });
-UX_STEP_NOCB(
-    ux_confirm_single_flow_3_step, 
-    bnnn_paging, 
-    {
-      .title = "Address",
-      .text = vars.tmp.fullAddress,
-    });
-UX_STEP_NOCB(
-    ux_confirm_single_flow_4_step, 
-    bnnn_paging, 
-    {
-      .title = "Fees",
-      .text = vars.tmp.feesAmount,
-    });
-UX_STEP_VALID(
-    ux_confirm_single_flow_5_step,
-    pb,
-    io_seproxyhal_touch_verify_ok(NULL), 
-    {
-      &C_icon_validate_14,
-      "Accept",
-    });
-UX_STEP_VALID(
-    ux_confirm_single_flow_6_step,
-    pb,
-    io_seproxyhal_touch_verify_cancel(NULL),
-    {
-      &C_icon_crossmark,
-      "Reject",
-    });
-// confirm_single: confirm output #x(feesAmount) / Amount: fullAmount / Address: fullAddress
-UX_FLOW(ux_confirm_single_flow,
-  &ux_confirm_single_flow_1_step,
-  &ux_confirm_single_flow_2_step,
-  &ux_confirm_single_flow_3_step,
-  &ux_confirm_single_flow_4_step,
-  &ux_confirm_single_flow_5_step,
-  &ux_confirm_single_flow_6_step
-);
-
-//////////////////////////////////////////////////////////////////////
-
-UX_STEP_NOCB(
-    ux_finalize_flow_1_step, 
-    pnn, 
-    {
-      &C_icon_eye,
-      "Review",
-      "transaction"
-    });
-UX_STEP_NOCB(
-    ux_finalize_flow_4_step,
-    bnnn_paging, 
-    {
-      .title = "Fees",
-      .text = vars.tmp.feesAmount,
-    });
-UX_STEP_VALID(
-    ux_finalize_flow_5_step,
-    pbb,
-    io_seproxyhal_touch_verify_ok(NULL),
-    {
-      &C_icon_validate_14,
-      "Accept",
-      "and send"
-    });
-UX_STEP_VALID(
-    ux_finalize_flow_6_step,
-    pb,
-    io_seproxyhal_touch_verify_cancel(NULL),
-    {
-      &C_icon_crossmark,
-      "Reject",
-    });
-// finalize: confirm transaction / Fees: feesAmount
-UX_FLOW(ux_finalize_flow,
-  &ux_finalize_flow_1_step,
-  &ux_finalize_flow_4_step,
-  &ux_finalize_flow_5_step,
-  &ux_finalize_flow_6_step
-);
-
-//////////////////////////////////////////////////////////////////////
-UX_STEP_NOCB(
-    ux_request_change_path_approval_flow_1_step,
-    pbb, 
-    {
-      &C_icon_eye,
-      "The change path",
-      "is unusual",
-    });
-UX_STEP_NOCB(
-    ux_request_change_path_approval_flow_2_step,
-    bnnn_paging, 
-    {
-      .title = "Change path",
-      .text = vars.tmp_warning.derivation_path,
-    });
-UX_STEP_VALID(
-    ux_request_change_path_approval_flow_3_step,
-    pbb,
-    io_seproxyhal_touch_display_cancel(NULL),
-    {
-      &C_icon_crossmark,
-      "Reject if you're",
-      "not sure",
-    });
-UX_STEP_VALID(
-    ux_request_change_path_approval_flow_4_step,
-    pb,
-    io_seproxyhal_touch_display_ok(NULL),
-    {
-      &C_icon_validate_14,
-      "Approve",
-    });
-
-UX_FLOW(ux_request_change_path_approval_flow,
-  &ux_request_change_path_approval_flow_1_step,
-  &ux_request_change_path_approval_flow_2_step,
-  &ux_request_change_path_approval_flow_3_step,
-  &ux_request_change_path_approval_flow_4_step
-);
-
-#endif // TARGET_NANOX || TARGET_NANOS2
-
-#ifndef TARGET_FATSTACKS
 // override point, but nothing more to do
 void io_seproxyhal_display(const bagl_element_t *element) {
     if ((element->component.type & (~BAGL_TYPE_FLAGS_MASK)) != BAGL_NONE) {
@@ -710,7 +111,6 @@ unsigned char io_event(unsigned char channel) {
     UNUSED(channel);
     // nothing done with the event, throw an error on the transport layer if
     // needed
-
     // can't have more than one tag in the reply, not supported yet.
     switch (G_io_seproxyhal_spi_buffer[0]) {
     case SEPROXYHAL_TAG_FINGER_EVENT:
@@ -762,423 +162,6 @@ unsigned char io_event(unsigned char channel) {
 
     // command has been processed, DO NOT reset the current APDU transport
     return 1;
-}
-
-uint8_t prepare_fees() {
-    if (btchip_context_D.transactionContext.relaxed) {
-        os_memmove(vars.tmp.feesAmount, "UNKNOWN", 7);
-        vars.tmp.feesAmount[7] = '\0';
-    } else {
-        unsigned char fees[8];
-        unsigned short textSize;
-        if (transaction_amount_sub_be(
-                fees, btchip_context_D.transactionContext.transactionAmount,
-                btchip_context_D.totalOutputAmount)) {
-            PRINTF("Fees: %.*H\n", 8, fees);
-            PRINTF("transactionAmount:: %.*H\n", 8, btchip_context_D.transactionContext.transactionAmount);
-            PRINTF("totalOutputAmount: %.*H\n", 8, btchip_context_D.totalOutputAmount);
-            PRINTF("Error : Fees not consistent");
-            goto error;
-        }
-        os_memmove(vars.tmp.feesAmount, btchip_context_D.shortCoinId,
-                   btchip_context_D.shortCoinIdLength);
-        vars.tmp.feesAmount[btchip_context_D.shortCoinIdLength] = ' ';
-        btchip_context_D.tmp =
-            (unsigned char *)(vars.tmp.feesAmount +
-                              btchip_context_D.shortCoinIdLength + 1);
-        textSize = btchip_convert_hex_amount_to_displayable(fees);
-        vars.tmp.feesAmount[textSize + btchip_context_D.shortCoinIdLength + 1] =
-            '\0';
-    }
-    return 1;
-error:
-    return 0;
-}
-
-uint8_t prepare_single_output() {
-    // TODO : special display for OP_RETURN
-    unsigned char amount[8];
-    char tmp[80];
-    unsigned int offset = 0;
-    unsigned char versionSize;
-    int addressOffset;
-    unsigned char address[22];
-    unsigned short version; // addr prefix, or net id
-    unsigned short textSize;
-    unsigned char script_version[2]; // Decred thing
-
-    vars.tmp.fullAddress[0] = '\0';
-    btchip_swap_bytes(amount, btchip_context_D.currentOutput + offset, 8);
-    offset += 8;
-
-    btchip_swap_bytes(script_version, btchip_context_D.currentOutput + offset, 2);
-    offset += 2;
-
-    PRINTF("amount: %.*H\n", 8, amount);
-
-    if (btchip_output_script_is_op_return(btchip_context_D.currentOutput +
-                                          offset)) {
-        strcpy(vars.tmp.fullAddress, "OP_RETURN");
-    }
-    else if (btchip_output_script_is_regular(btchip_context_D.currentOutput +
-                                             offset)) {
-        addressOffset = offset + 4;
-        version = btchip_context_D.payToAddressVersion;
-    }
-    else {
-        addressOffset = offset + 3;
-        version = btchip_context_D.payToScriptHashVersion;
-    }
-    if (vars.tmp.fullAddress[0] == 0) {
-
-        if (version > 255) {
-            versionSize = 2;
-            address[0] = (version >> 8);
-            address[1] = version;
-        } else {
-            versionSize = 1;
-            address[0] = version;
-        }
-        os_memmove(address + versionSize,
-                   btchip_context_D.currentOutput + addressOffset, 20);
-
-        // Prepare address
-        textSize = btchip_public_key_to_encoded_base58(
-                    address, 20 + versionSize, (unsigned char *)tmp,
-                    sizeof(tmp), version, 1);
-        tmp[textSize] = '\0';
-
-        strcpy(vars.tmp.fullAddress, tmp);
-    }
-
-    // Prepare amount
-
-    os_memmove(vars.tmp.fullAmount, btchip_context_D.shortCoinId,
-               btchip_context_D.shortCoinIdLength);
-    vars.tmp.fullAmount[btchip_context_D.shortCoinIdLength] = ' ';
-    btchip_context_D.tmp =
-        (unsigned char *)(vars.tmp.fullAmount +
-                          btchip_context_D.shortCoinIdLength + 1);
-    textSize = btchip_convert_hex_amount_to_displayable(amount);
-    vars.tmp.fullAmount[textSize + btchip_context_D.shortCoinIdLength + 1] =
-        '\0';
-
-    return 1;
-}
-
-uint8_t prepare_full_output(uint8_t checkOnly) {
-    PRINTF("prepare full output (check= %d):\n", checkOnly);
-    unsigned int offset = 0;
-    int numberOutputs;
-    int i;
-    unsigned int currentPos = 0;
-    unsigned char amount[8], totalOutputAmount[8], fees[8];
-    char tmp[80];
-    unsigned char outputPos = 0, changeFound = 0;
-    unsigned char script_version[2]; // Decred thing
-
-    if (btchip_context_D.transactionContext.relaxed &&
-        !btchip_context_D.transactionContext.consumeP2SH) {
-        if (!checkOnly) {
-            PRINTF("Error : Mixed inputs");
-        }
-        goto error;
-    }
-    if (btchip_context_D.transactionContext.consumeP2SH) {
-        if (checkOnly) {
-            goto error;
-        }
-        vars.tmp.fullAmount[0] = '\0';
-        vars.tmp.feesAmount[0] = '\0';
-        strcpy(vars.tmp.fullAddress, "P2SH");
-        return 1;
-    }
-    // Parse output, locate the change output location
-    os_memset(totalOutputAmount, 0, sizeof(totalOutputAmount));
-    numberOutputs = btchip_context_D.currentOutput[offset++];
-    PRINTF("%d outputs\n", numberOutputs);
-    if (numberOutputs > 3) {
-        if (!checkOnly) {
-            PRINTF("Error : Too many outputs");
-        }
-        goto error;
-    }
-    for (i = 0; i < numberOutputs; i++) {
-        unsigned char nullAmount = 1;
-        unsigned int j;
-        unsigned char isOpReturn, isP2sh;
-        unsigned char isOpCreate, isOpCall;
-
-        for (j = 0; j < 8; j++) {
-            if (btchip_context_D.currentOutput[offset + j] != 0) {
-                nullAmount = 0;
-                break;
-            }
-        }
-        btchip_swap_bytes(amount, btchip_context_D.currentOutput + offset, 8);
-        transaction_amount_add_be(totalOutputAmount, totalOutputAmount, amount);
-        offset += 8; // skip amount
-
-        btchip_swap_bytes(script_version, btchip_context_D.currentOutput + offset, 2);
-        offset += 2; // skip script_version
-
-        isOpReturn = btchip_output_script_is_op_return(
-            btchip_context_D.currentOutput + offset);
-        isP2sh = btchip_output_script_is_p2sh(btchip_context_D.currentOutput +
-                                              offset);
-        isOpCreate = btchip_output_script_is_op_create(
-            btchip_context_D.currentOutput + offset);
-        isOpCall = btchip_output_script_is_op_call(
-            btchip_context_D.currentOutput + offset);
-        PRINTF("REGULAR SCRIPT: %d\n", btchip_output_script_is_regular(btchip_context_D.currentOutput + offset));
-        // Always notify OP_RETURN to the user
-        if (nullAmount && isOpReturn) {
-            if (!checkOnly) {
-                PRINTF("Error : Unexpected OP_RETURN");
-            }
-            goto error;
-        }
-        if (!btchip_output_script_is_regular(btchip_context_D.currentOutput +
-                                             offset) &&
-            !isP2sh && !(nullAmount && isOpReturn) &&
-             (!isOpCreate && !isOpCall)) {
-            if (!checkOnly) {
-                PRINTF("Error : Unrecognized input script");
-            }
-            goto error;
-        } else if (!btchip_output_script_is_regular(
-                     btchip_context_D.currentOutput + offset) &&
-                 !isP2sh && !(nullAmount && isOpReturn)) {
-            if (!checkOnly) {
-                PRINTF("Error : Unrecognized input script");
-            }
-            goto error;
-        }
-        if (btchip_context_D.tmpCtx.output.changeInitialized && !isOpReturn) {
-            unsigned char addressOffset =
-                (isP2sh ? OUTPUT_SCRIPT_P2SH_PRE_LENGTH
-                                         : OUTPUT_SCRIPT_REGULAR_PRE_LENGTH);
-            if (os_memcmp(btchip_context_D.currentOutput + offset +
-                              addressOffset,
-                          btchip_context_D.tmpCtx.output.changeAddress + 1,
-                          20) == 0) {
-                if (changeFound) {
-                    if (!checkOnly) {
-                        PRINTF("Error : Multiple change output found");
-                    }
-                    goto error;
-                }
-                changeFound = 1;
-            } else {
-                // outputPos is the real output pointer (opposed to the change address output)
-                outputPos = currentPos;
-            }
-        }
-        offset += 1 + btchip_context_D.currentOutput[offset];
-        currentPos++;
-    }
-    if (btchip_context_D.tmpCtx.output.changeInitialized && !changeFound) {
-        if (!checkOnly) {
-            PRINTF("Error : change output not found");
-        }
-        goto error;
-    }
-    if (transaction_amount_sub_be(
-            fees, btchip_context_D.transactionContext.transactionAmount,
-            totalOutputAmount)) {
-        PRINTF("tx_amount: ", 8, btchip_context_D.transactionContext.transactionAmount);
-        PRINTF("total_amount: ", 8, totalOutputAmount);
-        if (!checkOnly) {
-            PRINTF("Error : Fees not consistent");
-        }
-        goto error;
-    }
-    if (!checkOnly) {
-        // Format validation message
-        currentPos = 0;
-        offset = 1;
-        btchip_context_D.tmp = (unsigned char *)tmp;
-        for (i = 0; i < numberOutputs; i++) {
-            if (!btchip_output_script_is_op_return(btchip_context_D.currentOutput + offset + 8 + 2)) {
-                unsigned char versionSize;
-                int addressOffset;
-                unsigned char address[22];
-                unsigned short version;
-
-                btchip_swap_bytes(amount, btchip_context_D.currentOutput + offset, 8);
-                offset += 8; // skip amount
-
-                btchip_swap_bytes(script_version, btchip_context_D.currentOutput + offset, 2);
-                offset += 2; // skip script_version
-
-                if (btchip_output_script_is_regular(
-                            btchip_context_D.currentOutput + offset)) {
-                    addressOffset = offset + 4;
-                    version = btchip_context_D.payToAddressVersion;
-                    } else {
-                    addressOffset = offset + 3;
-                    version = btchip_context_D.payToScriptHashVersion;
-                }
-                    if (version > 255) {
-                    versionSize = 2;
-                    address[0] = (version >> 8);
-                    address[1] = version;
-                    } else {
-                    versionSize = 1;
-                    address[0] = version;
-                }
-                os_memmove(address + versionSize,
-                            btchip_context_D.currentOutput + addressOffset,
-                            20);
-
-                // if we're processing the real output (not the change one)
-                if (currentPos == outputPos) {
-                    unsigned short textSize = 0;
-
-                    // Prepare address
-                    textSize = btchip_public_key_to_encoded_base58(
-                        address, 20 + versionSize, (unsigned char *)tmp,
-                        sizeof(tmp), version, 1);
-                    tmp[textSize] = '\0';
-
-                    strcpy(vars.tmp.fullAddress, tmp);
-
-                    // Prepare amount
-                    PRINTF("prepare amount\n");
-
-                    os_memmove(vars.tmp.fullAmount,
-                               btchip_context_D.shortCoinId,
-                               btchip_context_D.shortCoinIdLength);
-                    vars.tmp.fullAmount[btchip_context_D.shortCoinIdLength] =
-                        ' ';
-                    btchip_context_D.tmp =
-                        (unsigned char *)(vars.tmp.fullAmount +
-                                          btchip_context_D.shortCoinIdLength +
-                                          1);
-                    textSize = btchip_convert_hex_amount_to_displayable(amount);
-                    vars.tmp
-                        .fullAmount[textSize +
-                                    btchip_context_D.shortCoinIdLength + 1] =
-                        '\0';
-
-                    // prepare fee display
-                    PRINTF("prepare fee display\n");
-                    os_memmove(vars.tmp.feesAmount,
-                               btchip_context_D.shortCoinId,
-                               btchip_context_D.shortCoinIdLength);
-                    vars.tmp.feesAmount[btchip_context_D.shortCoinIdLength] =
-                        ' ';
-                    btchip_context_D.tmp =
-                        (unsigned char *)(vars.tmp.feesAmount +
-                                          btchip_context_D.shortCoinIdLength +
-                                          1);
-                    textSize = btchip_convert_hex_amount_to_displayable(fees);
-                    vars.tmp
-                        .feesAmount[textSize +
-                                    btchip_context_D.shortCoinIdLength + 1] =
-                        '\0';
-                    break;
-                }
-            } else {
-                // amount + version
-                offset += 8 + 2;
-            }
-            offset += 1 + btchip_context_D.currentOutput[offset];
-            currentPos++;
-        }
-    }
-    return 1;
-error:
-    return 0;
-}
-
-#define HASH_LENGTH 4
-uint8_t prepare_message_signature() {
-    uint8_t buffer[32];
-
-    cx_hash(&btchip_context_D.transactionHashWitness.header, CX_LAST,
-            vars.tmp.fullAmount, 0, buffer, 32);
-
-    snprintf(vars.tmp.fullAddress, sizeof(vars.tmp.fullAddress), "%.*H...%.*H",
-             8, buffer, 8, buffer + 32 - 8);
-    return 1;
-}
-
-unsigned int btchip_bagl_confirm_full_output() {
-    if (!prepare_full_output(0)) {
-        return 0;
-    }
-
-#if defined(TARGET_NANOS)
-    ux_step = 0;
-    ux_step_count = 4;
-    UX_DISPLAY(ui_verify_nanos, ui_verify_prepro);
-#elif defined(TARGET_NANOX) || defined(TARGET_NANOS2)
-    ux_flow_init(0, ux_confirm_full_flow, NULL);
-#endif // TARGET_NANOX || TARGET_NANOS2
-    return 1;
-}
-
-unsigned int btchip_bagl_confirm_single_output() {
-// TODO : remove when supporting multi output
-    if (!prepare_single_output()) {
-        return 0;
-    }
-
-    snprintf(vars.tmp.feesAmount, sizeof(vars.tmp.feesAmount), "output #%d",
-             btchip_context_D.totalOutputs - btchip_context_D.remainingOutputs +
-                 1);
-
-#if defined(TARGET_NANOS)
-    ux_step = 0;
-    ux_step_count = 3;
-    UX_DISPLAY(ui_verify_output_nanos, ui_verify_output_prepro);
-#elif defined(TARGET_NANOX) || defined(TARGET_NANOS2)
-    ux_flow_init(0, ux_confirm_single_flow, NULL);
-#endif // TARGET_NANOX || TARGET_NANOS2
-    return 1;
-}
-
-unsigned int btchip_bagl_finalize_tx() {
-    if (!prepare_fees()) {
-        return 0;
-    }
-
-#if defined(TARGET_NANOS)
-    ux_step = 0;
-    ux_step_count = 2;
-    UX_DISPLAY(ui_finalize_nanos, ui_finalize_prepro);
-#elif defined(TARGET_NANOX) || defined(TARGET_NANOS2)
-    ux_flow_init(0, ux_finalize_flow, NULL);
-#endif // TARGET_NANOX
-    return 1;
-}
-
-void btchip_bagl_confirm_message_signature() {
-    if (!prepare_message_signature()) {
-        return;
-    }
-
-#if defined(TARGET_NANOS)
-    ux_step = 0;
-    ux_step_count = 2;
-    UX_DISPLAY(ui_verify_message_signature_nanos, ui_verify_message_prepro);
-#elif defined(TARGET_NANOX) || defined(TARGET_NANOS2)
-    ux_flow_init(0, ux_sign_flow, NULL);
-#endif // TARGET_NANOX
-}
-
-void btchip_bagl_request_change_path_approval(unsigned char* change_path)
-{
-    bip32_print_path(change_path, vars.tmp_warning.derivation_path, MAX_DERIV_PATH_ASCII_LENGTH);
-#if defined(TARGET_NANOS)
-    // append and prepend a white space to the address
-    ux_step = 0;
-    ux_step_count = 4;
-    UX_DISPLAY(ui_request_change_path_approval_nanos, ui_request_change_path_approval_nanos_prepro);
-#elif defined(TARGET_NANOX) || defined(TARGET_NANOS2)
-    ux_flow_init(0, ux_request_change_path_approval_flow, NULL);
-#endif // #if TARGET_ID
 }
 
 void app_exit(void) {
